+++ SLIP_LU/Source/SLIP_create_mpfr_array.c
@@ -14,7 +14,7 @@
 
 mpfr_t* SLIP_create_mpfr_array
 (
-    int32_t n,     // size of the array
+    int64_t n,     // size of the array
     SLIP_options *option// command options containing the prec for mpfr
 )
 {
@@ -32,7 +32,7 @@ mpfr_t* SLIP_create_mpfr_array
 
     mpfr_t* x = (mpfr_t*) SLIP_calloc(n, sizeof(mpfr_t));
     if (!x) {return NULL;}
-    for (int32_t i = 0; i < n; i++)
+    for (int64_t i = 0; i < n; i++)
     {
         if (SLIP_mpfr_init2(x[i], option->prec) != SLIP_OK)
         {
